Move CoffeeMachineHandler and CoffeeMachineChannelSelector to internal package
